separate display handler

+++ controller/workout.cpp
@@ -22,18 +22,18 @@ public:
     return other(id);
   }
 
-  void getTimersById(int* output, uint8_t id) {
+  void getTimersById(short* output, uint8_t id) {
     char* details = getById(id);
 
     //Serial.println(details);
 
-    int multiplerItems[] = { 0, 0, 0, 0, 0 };
+    short multiplerItems[] = { 0, 0, 0, 0, 0 };
     bool inMultiplier = false;
     uint8_t outIndex = 0;
     uint8_t multiplerIndex = 0;
     uint8_t numericCounter = 0;
     uint8_t multiplier = 0;
-    for (int i = 0; i < strlen(details); i++) {
+    for (uint8_t i = 0; i < strlen(details); i++) {
       uint8_t charValue = (uint8_t)details[i];
       if (48 <= charValue && charValue <= 57) {  //between 0 and 9
         uint8_t numValue = charValue - 48;
@@ -68,8 +68,8 @@ public:
       } else if (charValue == 93) {  // ]
         inMultiplier = false;
         numericCounter = 0;
-        for (int i = 0; i < multiplerItems[0]; i++) {
-          for (int j = 1; j < 5; j++) {
+        for (uint8_t i = 0; i < multiplerItems[0]; i++) {
+          for (uint8_t j = 1; j < 5; j++) {
             if (multiplerItems[j] != 0) {
               output[outIndex] = multiplerItems[j];
               outIndex++;
@@ -87,7 +87,7 @@ public:
         numericCounter = 0;
       }
     }
-    for (int i = 0; i < 20; i++) {
+    for (uint8_t i = 0; i < 20; i++) {
       output[i] = output[i] * 60;
     }
   }
